+++ app/Http/Controllers/Backend/TransactionController.php
@@ -75,6 +75,11 @@ class TransactionController extends Controller
      */
     public function destroy(string $id)
     {
+        // Cek apakah user adalah operator
+        if (auth()->check() && auth()->user()->role === 'operator') {
+            return redirect()->route('panel.transaction.index')->with('error', 'Owner tidak diizinkan menghapus transaksi.');
+        }
+
         $getTransaction = Transaction::where('uuid', $id)->firstOrFail();
 
         $this->fileService->delete($getTransaction->file);
